ODP-2724: Improve some code, remove external sed command
This migrates away from using an external process to change versions within files, to
using pure rust instead so that behaviour is always the same, no matter
the platform.

change_release_version is now wrapped in tokio::task::spawn_blocking
so that the async runtime knows that this is a function that could get
stuck/block for a while (external process calling).

Slack notifications for version changes have been added as well.