ODP-5234: Add a repository backup blacklist
This adds the ability to blacklist certain repositories for a backup operation
in the git-manage.toml configuration file. If `--blacklist` is passed,
any repository in that list will not be backed up. This is particularly
useful for repositories with few/no changes from upstream that are
large (like Clickhouse).

This also fixes a major performance regression I introduced in
8ad706c969efb166a9d3092124e198647e29db0c where the maximum number
of parallel jobs was always set to 1.