ODP-5581: Add fetching branches/tags and filtering them by regex
This adds a 'show' command to both the branch and tag commands so
that you can display them for a specific repository, or all selected ones.
This can be used to search for branches/tags, but the functionality was
originally concieved of for improved mpack support. IE using the filte
to get all tags from acceldata-io/ambari-mpacks that end in an ODP version
like '3\.2\.3\.4-2-tag$', which would then give you a list of all mpacks for
that release which could then be used to download them as tarballs to an
output folder on a machine or the mirror server (a pending feature
request).
I've tested the functionality and updated the corresponding ticket with
some examples of their output.