ODP-5581: Cache regexes in a thread safe way, clean up regexes & branch/file modification
This fixes the regexes being compiled once and reused globally.
Now (many of) the regexes are compiled and stored in a Hash Map with a Mutex.

This also modifies the file modification code that was used for modifying branches,
as well as fixing the previous workaround for the temp_dir handling when doing a dry run,
which involved changing to a different library. But now both the dry run and real operations
paths resolve to a PathBuf; if it's a dry run, the path is shown in the terminal and not deleted.

I've also modified modify_branch's spawn_blocking block to return an enum instead of using
untyped strings to pass messages back to the function. This should make it easier to understand
and less error prone.