ODP-5611: Set status code to 0 if there are less failures than the number of repos
This changes the behaviour of exiting. Previously, if any error was returned when
managing multiple repositories, the exit code would be set to 1. This caused issues
for the Jenkins pipeline, where it interpreted that as the whole pipeline failing.

Doing this also allows you to do some command and then use '&&' if at least one
repository succeeds. There are lots of situations where there is a degree of expected
failures - not all components have a release or a previous release for some command,
or some components don't have a particular branch or tag.

This also changes the status code to display the number of failures, up to 255.
Status codes above 255 can cause issues with some cli tools.