ODP-5582: Add configurable prefix for tarball downloading
When downloading branches or tags across multiple repositories,
it's possible that multiple repos have the same tag name. If they did,
the downloaded tarball would get clobbered by the next downloaded tarball.

This adds a configurable prefix that, if the branch/tag begins with that,
will add the repository name at the beginning of the filename to prevent
overwriting other files that have already been downloaded in this
invocation of the program.